update time evolution and ratio production
