+++ src/produce_ratio.py
@@ -7,106 +7,6 @@ import time
 
 from find_range import find_run_range
 
-def data_hists(rdf, hist_config, bins):
-    hd = {}
-    hd["int_lumi"] = rdf.Mean("int_lumi").GetValue()
-    hd["min_run"] = rdf.Min("min_run").GetValue()
-    hd["max_run"] = rdf.Max("max_run").GetValue()
-    for hist in hist_config:
-        name = hist_config[hist]["name"]
-        title = hist_config[hist]["title"]
-        if hist_config[hist]["type"] == "Histo1D":
-            x_bins = hist_config[hist]["x_bins"]
-            x_val = hist_config[hist]["x_val"]
-            cut = hist_config[hist].get("cut")
-            if cut:
-                hd[hist] = rdf.Filter(cut).Histo1D((name, title, bins[x_bins]["n"], bins[x_bins]["bins"]),
-                        x_val, "weight")
-            else:
-                hd[hist] = rdf.Histo1D((name, title, bins[x_bins]["n"], bins[x_bins]["bins"]),
-                    x_val, "weight")
-        elif hist_config[hist]["type"] == "Profile1D":
-            x_bins = hist_config[hist]["x_bins"]
-            x_val = hist_config[hist]["x_val"]
-            y_val = hist_config[hist]["y_val"]
-            cut = hist_config[hist].get("cut")
-            if cut:
-                hd[hist] = rdf.Filter(cut).Profile1D((name, title, bins[x_bins]["n"], bins[x_bins]["bins"]),
-                        x_val, y_val, "weight")
-            else:
-                hd[hist] = rdf.Profile1D((name, title, bins[x_bins]["n"], bins[x_bins]["bins"]),
-                        x_val, y_val, "weight")
-    return hd
-
-def lumi_data(rdf, hist_config):
-    ld = {}
-    ld["int_lumi"] = rdf.Mean("int_lumi")
-    ld["min_run"] = rdf.Min("min_run")
-    ld["max_run"] = rdf.Max("max_run")
-    for hist in hist_config:
-        x_val = hist_config[hist]["x_val"]
-        y_val = hist_config[hist]["y_val"]
-        cut = hist_config[hist].get("cut")
-        if cut:
-            ld[hist] = rdf.Filter(cut).Stats(y_val, "weight")
-        else:
-            ld[hist] = rdf.Stats(y_val, "weight")
-    return ld
-
-
-def produce_cumulative(lds, lm, hist_config, bins):
-    handles = [lm[hist] for hist in hist_config]
-    for ld in lds:
-        for hist in hist_config:
-            handles.append(ld[hist])
-    ROOT.RDF.RunGraphs(handles)
-
-    clumi = 0.0
-    lumi = 0.0
-    lumi_bins = [0.0]
-    for ld in lds:
-        lumi = ld["int_lumi"].GetValue()
-        clumi += lumi
-        lumi_bins.append(clumi)
-
-    hs = {}
-    hs["min_runs"] = ROOT.TH1D(f"min_runs", "min_runs", len(lumi_bins)-1, np.array(lumi_bins))
-    hs["max_runs"] = ROOT.TH1D(f"max_runs", "max_runs", len(lumi_bins)-1, np.array(lumi_bins))
-    for hist in hist_config:
-        name = hist_config[hist]["name"]
-        title = f"{hist};Cumulative luminosity (fb^{{-1}});<Data/MC>"
-
-        hs[hist] = ROOT.TH1D(f"{name}_cumulative", title, len(lumi_bins)-1,
-                np.array(lumi_bins))
-
-    print("Filling cumulative histograms")
-    start_cumulative = time.time()
-    for i, ld in enumerate(lds):
-        int_lumi = ld["int_lumi"].GetValue()
-        min_run = ld["min_run"].GetValue()
-        max_run = ld["max_run"].GetValue()
-
-        hs["min_runs"].SetBinContent(i+1, min_run)
-        hs["max_runs"].SetBinContent(i+1, max_run)
-        for hist in hist_config:
-            start_ratio = time.time()
-            md = ld[hist].GetValue().GetMean()
-            ed = ld[hist].GetValue().GetMeanErr()
-            mm = lm[hist].GetValue().GetMean()
-            em = lm[hist].GetValue().GetMeanErr()
-
-            if mm > 0.0 and md > 0.0:
-                ratio = md / mm
-                ratio_err = np.abs(ratio)*np.sqrt((ed/md)**2+(em/mm)**2)
-                hs[hist].SetBinContent(i+1, ratio)
-                hs[hist].SetBinError(i+1, ratio_err)
-    print(f"Done filling cumulative histograms (took {time.time() - start_cumulative} s)")
-    hists = [hs["min_runs"], hs["max_runs"]]
-    for hist in hist_config:
-        hists.append(hs[hist])
-
-    return hists
-
 def produce_ratio(rdf_numerator, h_denominator, hist_config, bins, i=None):
     name = hist_config["name"]
     title = hist_config["title"]
@@ -204,12 +104,6 @@ def run(args):
             chain_data.Add(file)
             chain_runs.Add(file)
             rdf = ROOT.RDataFrame("Events", file)
-            if args.cumulative_lumi:
-                # Create data histograms here early so they do not need to be
-                # generated multiple times in produce_cumulative
-                ld = lumi_data(rdf, hist_config)
-                lds.append(ld)
-                print(f"{j+1}/{len(group)} lumi chunks recorded")
 
         rdf_data = ROOT.RDataFrame(chain_data)
         rdf_runs = ROOT.RDataFrame(chain_runs)
@@ -228,14 +122,11 @@ def run(args):
             output_path = f"{args.out}/J4PRatio_runs{min_run}to{max_run}_vs_{args.mc_tag}.root"
 
         file_ratio = ROOT.TFile.Open(f"{output_path}", "RECREATE")
-        if args.cumulative_lumi:
-            hs = produce_cumulative(lds, lm, hist_config, bins)
-            for h in hs:
-                h.Write()
-        else:
-            for hist in hist_config:
-                h = produce_ratio(rdf_data, hm[hist], hist_config[hist], bins)
-                h.Write()
+
+        for hist in hist_config:
+            h = produce_ratio(rdf_data, hm[hist], hist_config[hist], bins)
+            h.Write()
+
         file_ratio.Close()
 
         if len(groups) > 1:
